issue #193: publish bigann and gist1m as "custom" datasets on GCS
Add `StandardDatasetPublisher`, a tool that stages the native BIGANN or
GIST1M files in the flat layout VectorBench custom mode expects,
generates the descriptor JSON, and optionally invokes
`push_dataset_gcs.sh` to upload to a `gs://` path. Motivation: let
cloud runs consume the standard datasets from `gs://herddb-datasets`
rather than depending on the TEXMEX FTP mirror.

Supporting loader changes so custom mode can actually consume the
uploads:

- DatasetDescriptor gains a `format` field ("fvecs" or "bvecs",
  defaulting to "fvecs" for back-compat). DatasetLoader now resolves
  the base/query format from the descriptor when preset==CUSTOM, so
  BIGANN's `.bvecs` files are readable under `--dataset custom`.
- `totalVectors`, `numQueries`, `Config.numRows`, and `Config.resumeFrom`
  are widened to `long`; `streamBaseVectors(...)` and `skipVecsInStream`
  take `long` too. 1 B BIGANN already exceeds INT_MAX when combined,
  and 20 B-class corpora are on the roadmap.
- BIGANN's nested `bigann_gnd/idx_1000000000.ivecs` is flattened to
  `bigann_groundtruth.ivecs` in the staging dir so it sits next to the
  descriptor.

Tests cover the descriptor+loader round-trip for both FVECS and BVECS
without touching the network, plus the `format`-missing back-compat
case.

+++ vector-testings/src/main/java/herddb/vectortesting/BenchOutput.java
@@ -413,7 +413,7 @@ abstract class BenchOutput {
             m.put("jdbc_url", config.jdbcUrl);
             m.put("dataset", config.dataset.name());
             m.put("table", config.tableName);
-            m.put("rows", (long) config.numRows);
+            m.put("rows", config.numRows);
             m.put("ingest_threads", config.ingestThreads);
             m.put("batch_size", config.batchSize);
             m.put("query_threads", config.queryThreads);
